refactor(breadth): rename research/breadth → breadth (Phase R)
Production daily-pipeline code (breadth/) was misnamed under research/.
After the maunakea/rainier split, research/ contains research-only modules
(bandit, evaluator, rewards, providers, etc.) that will leave for maunakea
in Phase 2. The breadth substrate stays in rainier and gets promoted to
src/rainier/breadth/ to reflect its production role.

Mechanical-only changes:
- git mv src/rainier/research/breadth/{__init__,loaders,ml_builders,ranks,
  registry,universe_loader}.py → src/rainier/breadth/
- git mv tests/research/test_thematic_{labels,ml_builders,registry,loaders,
  universe_log,ranks}.py → tests/breadth/
- Rewrite `rainier.research.breadth` → `rainier.breadth` across cli.py +
  breadth/__init__.py + the 6 moved test files
- tests/breadth/__init__.py added (empty, mirrors tests/research/__init__.py)

Acceptance checks:
- `rg "rainier\.research\.breadth"` returns zero in src/ tests/ scripts/
- `uv run pytest tests/ -q` → 990 passed / 7 skipped
- `uv run ruff check src/ tests/` clean
- `uv run rainier thematic --help` shows full subcommand tree
- Byte-identical daily-pipeline render: SHA256 of
  docs/thematic-ranks-2026-05-01.html unchanged across rename
  (292bf54f...22801 baseline == 292bf54f...22801 post-rename)

Out of scope (Phase 2 / future PRs):
- research/{bandit,evaluator,rewards,providers,cost_pilot,survivorship,...}
  stay until maunakea Phase 2 ports them.
- research/macro_context_loader.py move happens in a separate task.

External users: rewrite imports with
    sed -i 's/rainier\.research\.breadth/rainier.breadth/g'
